refactor: fix all entities

+++ src/service/workTemplate/createWorkTemplateService.ts
@@ -2,15 +2,16 @@ import { getRepository } from 'typeorm';
 
 import { DayTemplateInput } from 'interfaces/workTemplate';
 import { DoctorWorkTemplate } from 'typeorm/entities/doctorWorkTemplate';
+import { Doctor } from 'typeorm/entities/doctor';
 
 export const createWorkTemplateService = async (doctorId: string, templates: DayTemplateInput[]) => {
   const doctorWorkTemplateRepository = getRepository(DoctorWorkTemplate);
 
-  await doctorWorkTemplateRepository.delete({ doctorId });
+  await doctorWorkTemplateRepository.delete({ doctor: { user_id: doctorId } });
 
   const newTemplates = templates.map((t) => {
     const template = new DoctorWorkTemplate();
-    template.doctorId = doctorId;
+    template.doctor = { user_id: doctorId } as Doctor;
     template.dayOfWeek = t.dayOfWeek;
     template.isAvailable = t.isAvailable;
     template.morningStartTime = t.morningStartTime || '00:00';
